Refactor 2.3.2 baseline (no ORM support)
