--- Bench-Kitura-TechEmpower/2.3.2/Sources/TechEmpowerKueryMustache/Fortune+Mustache.swift
@@ -1,7 +1,0 @@
-import Mustache
-
-extension Fortune: MustacheBoxable {
-  var mustacheBox: MustacheBox {
-    return Box(["id": self.id, "message": self.message])
-  }
-}
