+++ Bench-Kitura-TechEmpower/2.3.2/Sources/TechEmpowerPostgresMustache/Fortune+Mustache.swift
@@ -1,5 +1,5 @@
 /*
- * Copyright IBM Corporation 2016
+ * Copyright IBM Corporation 2018
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-public enum AppError: Error {
-    case DBError(String, query: String)
-    case MongoError(String)
-    case DataFormatError(String)
-    case OtherError(String)
+import Mustache
+import TechEmpowerCommon
+
+extension Fortune: MustacheBoxable {
+  public var mustacheBox: MustacheBox {
+    return Box(["id": self.id, "message": self.message])
+  }
 }
